==> notebooks/02_Data_Cleaning_and_Preprocessing.ipynb <==
import pandas as pd
import numpy as np

#Load the dataset
df_raw=pd.read_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\raw\application_train.csv")
df_raw.head()

#drop ID column
df_model_base=df_raw.drop('SK_ID_CURR',axis=1)
df_model_base.info()

#import train_test_split
from sklearn.model_selection import train_test_split

x=df_model_base.drop('TARGET',axis=1)
y=df_model_base['TARGET']

x_train,x_test,y_train,y_test=train_test_split(x,y,test_size=0.2,stratify=y)

x_train.shape

x_test.shape

y_test.value_counts(normalize=True)

y_train.value_counts(normalize=True)

#Does any column contain post-loan performance info?
pd.set_option('display.max_columns', 10000)
#Display all the rows
pd.set_option('display.max_rows', None)

x_train.head()

missing=x_train.isnull().mean()
missing

missing.sort_values(ascending=False)

#GET THE COLUMNS WITH MORE THAN 60% MISSING VALUES
more_than_60=missing[missing>0.6].index.tolist()

#CREATE FLAGS FOR THOSE COLS
for col in more_than_60:
    x_train[col+"_missing_flag"]=x_train[col].isnull().astype(int)
    x_test[col+"missing_flag"]=x_test[col].isnull().astype(int)

len(more_than_60)

#DO IMPUTATION FOR THOSE COLS
def imputation_for_missing_values(df1,df2,cols):
    for col in cols:
        if df1[col].dtype=='object':
            df1.loc[(df1[col].isnull()),col]='Missing'
            df2.loc[(df2[col].isnull()),col]='Missing'
        else:
            df1.loc[(df1[col].isnull()),col]=df1[col].median()
            df2.loc[(df2[col].isnull()),col]=df1[col].median()

all_missing_cols=missing[missing>0].index.tolist()

imputation_for_missing_values(x_train,x_test,all_missing_cols)

#check if any null values are present in x_train x_test
print(x_train.isnull().sum())
print(x_test.isnull().sum())

x_test.isnull().sum()

#missing flag counts
flag_cols = [col for col in x_train.columns if col.endswith("_missing_flag")]

print("Number of missing flag columns:", len(flag_cols))
print("Missing flag columns:", flag_cols[:5])  # show first 5

sample_col = flag_cols[0].replace("_missing_flag", "")

check_df = x_train[[sample_col, sample_col + "_missing_flag"]].head(10)
print(check_df)

check_df.loc[(check_df['OWN_CAR_AGE_missing_flag']==1),:]

print("Total NaNs in X_train:", x_train.isnull().sum().sum())
print("Total NaNs in X_test:", x_test.isnull().sum().sum())


# Export the cleaned datasets

x_train.to_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\x_train_clean.csv",index=False)
x_test.to_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\x_test_clean.csv",index=False)
y_train.to_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\y_train.csv",index=False)
y_test.to_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\y_test.csv",index=False)
