==> notebooks/03_Variable_Selection_WOE_IV.ipynb <==
# LOAD THE TRAIN AND TEST DATASETS

import pandas as pd
import numpy as np

x_train=pd.read_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\x_train_clean.csv")
x_test=pd.read_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\x_test_clean.csv")

#LOAD Y_TRAIN AND Y_TEST
y_train=pd.read_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\y_train.csv")
y_test=pd.read_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\y_test.csv")

print(x_train.shape)
print(x_test.shape)
print(y_train.shape)
print(y_test.shape)

print(x_train.head())

## MERGE X_TRAIN AND Y_TRAIN
BECAUSE WOE BINNING NEEDS TARGET VARIABLE TO CALCULATE WOE

train_data=x_train.copy(deep=True)

train_data['TARGET']=y_train

#shape of train_Data
train_data.shape

train_data.columns

train_data.isnull().sum().sum()

train_data['TARGET'].value_counts()

# SEPARATE NUMERICAL COLS CATEGORICAL COLS AND MISSING FLAG COLS

pd.set_option('display.max_info_columns',100000)

train_data.info()

numerical_cols=train_data.select_dtypes(include=['int64','float64']).columns.tolist()
print(numerical_cols)

#remove target from numerical cols and missing flag cols
numerical_cols.remove('TARGET')

#remove missing flag cols
missing_flag_cols=[i for i in numerical_cols if '_missing_flag' in i]
# numerical_cols=[i for i in numerical_cols if '_missing_flag' in i==False]
# print(missing_flag_cols)

numerical_cols=[i for i in numerical_cols if ('_missing_flag' not in i)]

numerical_cols

#get categorical column
categorical_cols=train_data.select_dtypes(include='object').columns.tolist()

categorical_cols

print("Numerical cols ", len(numerical_cols))
print("Categorical cols ", len(categorical_cols))
print("Flag cols ", len(missing_flag_cols))

# CREATE BIN FOR ONE NUMERICAL VARIABLE
EXT_SOURCE_2

train_data['EXT_SOURCE_2_BIN']=pd.qcut(train_data['EXT_SOURCE_2'],q=10)

EXT_SOURCE_2_BIN_CROSSTAB=pd.crosstab(index=train_data['EXT_SOURCE_2_BIN'],columns=train_data['TARGET'])

EXT_SOURCE_2_BIN_CROSSTAB

#total defaults and non defaults
EXT_SOURCE_2_BIN_CROSSTAB.apply(lambda x: x.sum(), axis=0)

## Create two new columns:
dist_non_default
dist_default

EXT_SOURCE_2_BIN_CROSSTAB[['dist_non_default','dist_default']]=EXT_SOURCE_2_BIN_CROSSTAB.apply(lambda x: x/x.sum(), axis=0)
EXT_SOURCE_2_BIN_CROSSTAB

#CALCULATE WOE
EXT_SOURCE_2_BIN_CROSSTAB['WOE']=np.log(EXT_SOURCE_2_BIN_CROSSTAB['dist_non_default']/EXT_SOURCE_2_BIN_CROSSTAB['dist_default'])
EXT_SOURCE_2_BIN_CROSSTAB

#CALCULATE IV
EXT_SOURCE_2_BIN_CROSSTAB['IV_COMPONENT']=(EXT_SOURCE_2_BIN_CROSSTAB['dist_non_default']-EXT_SOURCE_2_BIN_CROSSTAB['dist_default'])*EXT_SOURCE_2_BIN_CROSSTAB['WOE']
EXT_SOURCE_2_BIN_CROSSTAB

total_iv=EXT_SOURCE_2_BIN_CROSSTAB['IV_COMPONENT'].sum()
total_iv

# CREATING A BINNING FUNCTION

def create_quantile_bins(dataframe,variable, n_bins=10):
    binning_series, bin_edges=pd.qcut(dataframe[variable],q=n_bins,duplicates='drop',retbins=True)
    return binning_series, bin_edges
    

#Testing above function on ext_source_2
train_data["EXT_SOURCE_2_BIN"], ext2_edges=create_quantile_bins(train_data,'EXT_SOURCE_2')

ext2_edges

train_data["EXT_SOURCE_2_BIN"]

# CREATE WOE IV FUNCTION

def calculate_woe_iv(dataframe, binned_col_name, target_variable):
    
    # Step 1: Crosstab
    cross_tab = pd.crosstab(
        index=dataframe[binned_col_name],
        columns=dataframe[target_variable]
    )
    
    # Ensure both 0 and 1 columns exist
    if 0 not in cross_tab.columns:
        cross_tab[0] = 0
    if 1 not in cross_tab.columns:
        cross_tab[1] = 0

    # Step 2: Total goods and bads
    total_good = cross_tab[0].sum()
    total_bad = cross_tab[1].sum()
    
    # Step 3: Distribution
    cross_tab['%good'] = cross_tab[0] / total_good
    cross_tab['%bad'] = cross_tab[1] / total_bad
    
    # Step 4: Replace 0 with small value to avoid division by zero
    epsilon = 1e-6
    cross_tab['%good'] = cross_tab['%good'].replace(0, epsilon)
    cross_tab['%bad'] = cross_tab['%bad'].replace(0, epsilon)
    
    # Step 5: WOE
    cross_tab['WOE'] = np.log(cross_tab['%good'] / cross_tab['%bad'])
    
    # Step 6: IV component
    cross_tab['IV_COMPONENT'] = (
        (cross_tab['%good'] - cross_tab['%bad']) * cross_tab['WOE']
    )
    
    # Step 7: Total IV
    total_iv = cross_tab['IV_COMPONENT'].sum()
    
    return cross_tab, total_iv


ext_source_2_ct, ext_source_2_iv=calculate_woe_iv(train_data,"EXT_SOURCE_2_BIN","TARGET")

ext_source_2_ct

ext_source_2_iv

# LOOP OVER ALL NUMERIC COLS AND RUN BINNING AND WOE functions

#CREATE DICTIONARIES TO STORE
#IV VALUES OF COLS
#WOE TABLES FOR EACH COLS BINS
#BIN EDGES FOR COLS
woe_tables = {}
bin_edges_dict = {}
iv_dict = {}

# NOW WE LOOP OVER EACH COL AND RUN THE FUNCTS
for col in numerical_cols:
    try:
        binned_series,bin_edges=create_quantile_bins(train_data,col)
        temp_df=pd.DataFrame({col+"_BIN":binned_series,"TARGET":train_data["TARGET"]})
        woe_ct, total_iv=calculate_woe_iv(temp_df,col+"_BIN","TARGET")
        woe_tables[col]=woe_ct
        bin_edges_dict[col]=bin_edges
        iv_dict[col]=total_iv
    except Exception as e:
        print(f"{col} had some error {e}, skipping.")

# converting iv diction to dataframe
iv_summary_numeric = pd.DataFrame.from_dict(
    iv_dict,
    orient='index',
    columns=['IV']
).sort_values(by='IV', ascending=False)

iv_summary_numeric.head(10)


# CREATE WOE AND IVs FOR CATEGORICAL AND MISSING FLAG COLS

print(len(categorical_cols))
len(missing_flag_cols)

#combine both categorical and missing flag col as missing flag cols are categorical only
categorical_all=categorical_cols+missing_flag_cols
categorical_all

#Loop through each categorical col and create woe and iv(BINNING NOT REQUIRED HERE)
for col in categorical_all:
    try:
        woe_ct, total_iv = calculate_woe_iv(train_data, col, "TARGET")
        woe_tables[col]=woe_ct
        iv_dict[col]=total_iv
    except Exception as e:
        print(f"{col} variable has a an error {e}. Skipping")

iv_summary_all = pd.DataFrame.from_dict(
    iv_dict,
    orient='index',
    columns=['IV']
).sort_values(by='IV', ascending=False)

pd.set_option('display.max_rows',10000)
iv_summary_all.head(15)

# NOW VARIABLE ELIMINATION BASED ON IV

type(iv_summary_all)

#taking IV Cutoff as 0.05
selected_vars=iv_summary_all[iv_summary_all['IV']>=0.05].index.tolist()

len(selected_vars)

#Final DF after woe and iv
train_model_base=train_data[selected_vars + ["TARGET"]]

#EXPORT MODEL BASE TO PROCESSED DATA SET
train_model_base.to_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\train_model_base.csv",index=False)

# EXPORT THE DICTIONARIES AND SELECTED VARIABLES
1) SELECTED VARS
2) WOE TABLES
3) BIN EDGES DICTIONARY

import os

os.makedirs("models", exist_ok=True)

import pickle

model_artifacts = {
    "selected_vars": selected_vars,
    "woe_tables": woe_tables,
    "bin_edges_dict": bin_edges_dict
}

with open("models/model_artifacts.pkl", "wb") as f:
    pickle.dump(model_artifacts, f)
