==> notebooks/04_WOE_Transformation.ipynb <==
# LOAD ALL THE THINGS COMPUTED FROM NOTEBOOK 03

import pickle
with open("models/model_artifacts.pkl", "rb") as f:
    artifacts = pickle.load(f)

selected_vars = artifacts["selected_vars"]
woe_tables = artifacts["woe_tables"]
bin_edges_dict = artifacts["bin_edges_dict"]

import pandas as pd
import numpy as np

train_model_base=pd.read_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\train_model_base.csv")

# TRANSFORM VARIABLES TO THEIR WOEs

#checking how woe table looks like
woe_tables[selected_vars[0]]

#create separate function for numerical variables
def transform_numeric_to_woe(dataframe,variable):
    bin_edges=bin_edges_dict[variable]
    binned_series=pd.cut(dataframe[variable],bins=bin_edges,include_lowest=True)
    woe_mapping=woe_tables[variable]['WOE']
    transformed_series=binned_series.map(woe_mapping)
    return transformed_series

EXT_SOURCE_2_TRANSFORMED=transform_numeric_to_woe(train_model_base,"EXT_SOURCE_2")
EXT_SOURCE_2_TRANSFORMED.value_counts(dropna=False)

#FUNCTION FOR CATEGORICAL VARIABLE
def transform_categorical_to_woe(dataframe,variable):
    woe_mapping=woe_tables[variable]['WOE']
    transformed_series=dataframe[variable].map(woe_mapping).fillna(0)
    return transformed_series

# TEST THE ABOVE FUNCTION ON TEST_DATA

#import test data
x_test_clean=pd.read_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\x_test_clean.csv")

EXT_SOURCE_2_TRANSFORMED=transform_numeric_to_woe(x_test_clean,"EXT_SOURCE_2")
EXT_SOURCE_2_TRANSFORMED.value_counts(dropna=False)

# BULK TRANSFORMATION WOE

#create 2 dataframes 
x_train_woe=pd.DataFrame()
x_test_woe=pd.DataFrame()

train_model_base.info()

for col in selected_vars:
    if train_model_base[col].dtype=='object':
        x_train_woe[col]=transform_categorical_to_woe(train_model_base,col)
        x_test_woe[col]=transform_categorical_to_woe(x_test_clean,col)
    else:
        x_train_woe[col]=transform_numeric_to_woe(train_model_base,col)
        x_test_woe[col]=transform_numeric_to_woe(x_test_clean,col)

x_train_woe.head(15)

x_test_woe.head(15)

x_train_woe.isnull().sum().sum()

x_test_woe.isnull().sum().sum()

x_train_woe.shape

x_test_woe.shape

# EXPORT X_TRAIN_WOE AND X_TEST_WOE

x_train_woe.to_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\x_train_woe.csv",index=False)

x_test_woe.to_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\x_test_woe.csv",index=False)
