==> notebooks/05_SFA.ipynb <==
# LOAD THE DATASETS

import pandas as pd
import numpy as np

x_train_woe=pd.read_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\x_train_woe.csv")

y_train=pd.read_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\y_train.csv")

# BUILD THE LOGISTIC REGRESSEION MODEL FOR EACH VARIABLE

import statsmodels.api as sm
from sklearn.metrics import roc_auc_score

columns=x_train_woe.columns.tolist()

variable=[]
coef=[]
p_value=[]
auc=[]
somers_D=[]

for col in columns:
    x=sm.add_constant(x_train_woe[col])
    model=sm.Logit(y_train,x).fit()
    variable.append(col)
    coef.append(model.params[col])
    p_value.append(model.pvalues[col])
    y_pred_probab=model.predict(x)
    auc_score=roc_auc_score(y_train,y_pred_probab)
    auc.append(auc_score)
    somers_d_score=2*auc_score-1
    somers_D.append(somers_d_score)

sfa_summary=pd.DataFrame(
    {
        'Variable':variable,
        'Coeff':coef,
        'p-value':p_value,
        'auc':auc,
        "Somer's d":somers_D
    }
)

sfa_summary.sort_values("Somer's d",ascending=False)

x_train_woe.corr()

# Inferences from SFA
1) Coefficients of each variable is -ve because of WOE transormation of variable
2) p-value of each variable is negligible(<0.05) ie. significant
3) Strong variables (Somer's D > 0.25)
4) Medium Strength (0.15-0.20)
5) Weak(<0.12)

## We do not eliminate based any variables based on SFA as of now, because the signs are expected no instability p-values<0.05, somer's d>0.05
