==> notebooks/06_MFA.ipynb <==
# LOAD THE X_TRAIN_WOE AND Y_TRAIN DATASETS

import pandas as pd
import numpy as np

x_train_woe=pd.read_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\x_train_woe.csv")
y_train=pd.read_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\y_train.csv")

# FIT THE MODEL WITH ALL THE 10 VARIABLES

import statsmodels.api as sm

#add constant
x=sm.add_constant(x_train_woe)

#Fit the model
model=sm.Logit(y_train,x).fit()

model.summary()

**Inferences after fitting model with 10 variables**
1) Sign changed for DAYS_BIRTH, meaning variable is unstable
2) p-value for all the variables <0.05 ie. significant variables

## REFIT THE MODEL AFTER DROPPING DAYS_BIRTH

x_train_iter2=x_train_woe.drop("DAYS_BIRTH",axis=1)

x=sm.add_constant(x_train_iter2)

model_itr2=sm.Logit(y_train,x).fit()

model_itr2.summary()

**Inferences after fitting model with 9 variables(dropped DAYS_BIRTH)**
1) No sign flips, stable variables
2) p-value <0.05 significant variables

Now, we check multicolinearity

# CHECK FOR MULTICOLINEARITY WITHIN THE 9 REMAINING VARIABLES

def calculate_vif(dataframe, remove_cols=None):
    if remove_cols:
        dataframe.drop(remove_cols,axis=1,inplace=True)
    columns=dataframe.columns.tolist()
    for col in columns:
        y=dataframe[col]
        x=dataframe.drop(col,axis=1)
        x=sm.add_constant(x)
        model_linear=sm.OLS(y,x).fit()
        r_squ=model_linear.rsquared
        vif=1/(1-r_squ)
        print(f"VIF for {col} is : {vif}")

# for col in columns:
#     y=col
#     x=x_train_iter2.drop(col,axis=1)
#     x=sm.add_constant(x)
#     model_linear=sm.OLS(y,x).fit()
#     r_squ=model_linear.rsquared
#     vif=1/(1-squ)
#     print(f"VIF for {col} is : {vif}")

#calculate vif for x_train_iter_2
calculate_vif(x_train_iter2)

**Inference after calcuating VIF**
-No Multicolinearity
-VIF<5 for each variable 

# FINALIZE X_TRAIN AND X_TEST

x_train_final=x_train_iter2

x_test_woe=pd.read_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\x_test_woe.csv")

x_test_woe.head()

x_test_final=x_test_woe.drop("DAYS_BIRTH",axis=1)

# EXPORT X_TEST_FINAL AND X_TRAIN_FINAL

x_test_final.to_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\x_test_final.csv",index=False)
x_train_final.to_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\x_ttrain_final.csv",index=False)
