==> notebooks/07_Model_Validation_and_Performance.ipynb <==
# Load the datasets

import pandas as pd
import numpy as np

x_train_final=pd.read_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\x_train_final.csv")
x_test_final=pd.read_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\x_test_final.csv")

print(f"Train : {x_train_final.shape}")
print(f"Test : {x_test_final.shape}")

#load y_train and y_test
y_train=pd.read_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\y_train.csv")
y_test=pd.read_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\y_test.csv")

print(f"Train : {y_train.shape}")
print(f"Test : {y_test.shape}")

# FIT THE LOGIT MODEL

import statsmodels.api as sm

#ADD THE CONSTANT
x=sm.add_constant(x_train_final)

#fit the model
model=sm.Logit(y_train,x).fit()
model.summary()

# PREDICT PROBABILITIES

x_train_final=sm.add_constant(x_train_final)
x_test_final=sm.add_constant(x_test_final)

y_prob_train=model.predict(x_train_final)
y_prob_test=model.predict(x_test_final)

# CALCULATE AUC, SOMER'S D OF TRAIN AND TEST

from sklearn.metrics import roc_auc_score

auc_train=roc_auc_score(y_train,y_prob_train)
auc_test=roc_auc_score(y_test,y_prob_test)

print(f"Auc train : {auc_train}")
print(f"Auc test : {auc_test}")

# CALCULATE SOMER'S D OF TEST AND TRAIN
somers_d_train=2*auc_train-1
somers_d_test=2*auc_test-1

print(f"Somer's D train : {somers_d_train}")
print(f"Somer's D test : {somers_d_test}")

# CALCULATE KS-STATISTIC

y_prob_train = pd.Series(y_prob_train, name="pd")

y_train = y_train.squeeze()

ks_train=pd.DataFrame({"y_train":y_train,"pd":y_prob_train})

print("y_train shape:", y_train.shape)
print("y_prob_train shape:", y_prob_train.shape)

ks_train.sort_values("pd",ascending=False,inplace=True)

ks_train['decile']=pd.qcut(ks_train['pd'],q=10)

ks_train

ks_train=ks_train.groupby('decile',as_index=False).agg(
    total=('y_train', 'count'),
    bad=('y_train', 'sum'))

ks_train['good']=ks_train['total']-ks_train['bad']

ks_train=ks_train.sort_values("decile",ascending=False)
ks_train

ks_train['bad_cum_sum%']=ks_train['bad'].cumsum()/ks_train['bad'].sum()
ks_train['good_cum_sum%']=ks_train['good'].cumsum()/ks_train['good'].sum()

ks_train['ks']=ks_train['bad_cum_sum%']-ks_train['good_cum_sum%']

ks_stat=ks_train['ks'].max()

ks_stat

# CREATE A FUNCTION FOR KS STAT CALCULATION

def ks_stat_calculation(y,probability):
    y=y.squeeze()
    probability=pd.Series(probability).squeeze()
    df=pd.DataFrame({"y":y,"PD":probability})
    df.sort_values("PD",ascending=False,inplace=True)
    df['DECILE']=pd.qcut(df['PD'],q=10)
    df= df.groupby("DECILE",as_index=False).agg(total=('y','count'),bad=('y','sum'))
    df['good']=df['total']-df['bad']
    df['cum_sum_bad%']=df['bad'].cumsum()/df['bad'].sum()
    df['cum_sum_good%']=df['good'].cumsum()/df['good'].sum()
    df['ks']=abs(df['cum_sum_bad%']-df['cum_sum_good%'])
    ks_stat=df['ks'].max()
    return ks_stat

#test the function on y_train
print(f"KS TRAIN: {ks_stat_calculation(y_train,y_prob_train)}")

print(f"KS TEST: {ks_stat_calculation(y_test,y_prob_test)}")

# Inferences from Model Validation

|AUC|SOMERS D|KS STATS|STATUS|
|---|---|---|---|
|"0.75+"|"0.5+"|"35%"|"STRONG"|
|"0.7-0.75"|"0.4-0.5"|"30-35%"|"GOOD"|

1) AUC TRAIN ~ 0.737
2) AUC TEST ~ 0.731
3) SOMER'S D TRAIN~ 0.474
4) SOMER'S TEST~0.463
5) KS STAT TRAIN ~ 0.34
6) KS STATS TEST ~ 0.33

Train and test metrics are closely aligned, indicating good generalization and no evidence of overfitting.

# EXPORT MODEL COEFFICIENTS, MODEL SUMMARY, PREDICTED PD(TRAIN AND TEST) 

model.params

# Export coeeficients
model.params.to_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\final_model_coefficients.csv")

model.summary()

summary_table = pd.DataFrame({
    "Variable": model.params.index,
    "Coefficient": model.params.values,
    "Std_Error": model.bse.values,
    "z_value": model.tvalues.values,
    "p_value": model.pvalues.values
})

summary_table.to_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\model_summary.csv")

#Export predicted probabilities from train and test
y_prob_test.to_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\test_pd.csv")
y_prob_train.to_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\train_pd.csv")

## Author
Pranjal Saxena
