==> notebooks/08_Scorecard_Scaling_and_Points.ipynb <==
# LOAD THE COEFFICIENTS, SUMMARY AND PROBABILITIES

import pandas as pd
import numpy as np

final_model_coefficients=pd.read_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\final_model_coefficients.csv")
model_summary=pd.read_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\model_summary.csv")
test_pd=pd.read_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\test_pd.csv")
train_pd=pd.read_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\train_pd.csv")

x_train_final=pd.read_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\x_train_final.csv")

x_test_final=pd.read_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\x_test_final.csv")

# FIX THE SCORE DESIGN PARAMETERS
1) BASE SCORE: 600
2) PDO: 20
3) BASE ODDS: 50

base_odds=50
base_score=600
pdo=20

# CALCULATE THE FACTOR THAT NEEDS TO BE MULTIPLIED BY COEFFICIENTS AND WOE 
# FACTOR= PDO/ln(2)
factor=(20/np.log(2))
factor

# CALCUATE OFFSET
# OFFSET= BASE SCORE - FACTOR * ln(base odds)
offset=base_score- (factor*np.log(base_odds))
offset

# COMPUTE BASE POINTS (INTERCEPT CONTRIBUTION)

final_model_coefficients

intercept=final_model_coefficients.iloc[0,1]
intercept

base_points=offset-(factor*intercept)
base_points

# Calculate WOE Points for each bin 

#import the WOE tables
import pickle
with open("models/model_artifacts.pkl", "rb") as f:
    artifacts = pickle.load(f)

selected_vars = artifacts["selected_vars"]
woe_tables = artifacts["woe_tables"]
bin_edges_dict = artifacts["bin_edges_dict"]

woe_tables['EXT_SOURCE_2']

final_model_coefficients.rename(columns={'Unnamed: 0':'Variables','0':"Coeff"},inplace=True)

coef_var=final_model_coefficients[final_model_coefficients["Variables"]=='EXT_SOURCE_2']["Coeff"].values[0]
type(coef_var)

woe_tables['EXT_SOURCE_2']['BIN_PTS']=-1*coef_var*factor*woe_tables['EXT_SOURCE_2']['WOE']
woe_tables['EXT_SOURCE_2']

woe_bin_pts={}

def calculate_bin_pts(variable):
    df=woe_tables[variable]
    coeff_var=final_model_coefficients.loc[(final_model_coefficients["Variables"]==variable),"Coeff"].values[0]
    df["BIN_PTS"]=-factor*coeff_var*df['WOE']
    woe_bin_pts[variable]=df

calculate_bin_pts('EXT_SOURCE_3')
woe_bin_pts['EXT_SOURCE_3']

for col in x_train_final.columns.tolist():
    calculate_bin_pts(col)

# Build the final scorecard 

x_train_final.head()

test=x_train_final.iloc[[0],:]
test

test["EXT_SOURCE_3"].values[0]

abs(woe_bin_pts['EXT_SOURCE_3']['WOE']-test["EXT_SOURCE_3"].values[0])<1e-6


score_scorecard = base_points

for col in x_train_final.columns:
    beta = final_model_coefficients.loc[
        final_model_coefficients["Variables"] == col,
        "Coeff"
    ].values[0]
    
    woe_val = test[col]
    
    score_scorecard -= factor * beta * woe_val

print("Score (Scorecard method):", score_scorecard)

# compute logit manually
logit = intercept

for col in x_train_final.columns:
    beta = final_model_coefficients.loc[
        final_model_coefficients["Variables"] == col,
        "Coeff"
    ].values[0]
    
    logit += beta * test[col]

score_logistic = offset - factor * logit

print("Score (Logistic method):", score_logistic)

# CREATE A FUNCTION TO AUTOMATE SCORES FOR ALL VARIABLES

x_train_final.head()

def calculate_score(model_coefficients,base_points,dataframe,factor):
    score=base_points
    for col in dataframe.columns.tolist():
        beta=model_coefficients.loc[model_coefficients["Variables"]==col,"Coeff"].values[0]
        score-=factor*beta*dataframe[col]
    dataframe["SCORE"]=score
    return dataframe

x_train_scores=calculate_score(final_model_coefficients,base_points,x_train_final,factor)
x_train_scores.head()

x_test_scores=calculate_score(final_model_coefficients,base_points,x_test_final,factor)
x_test_scores.head()

x_train_scores.to_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\x_train_scores.csv",index=False)
x_test_scores.to_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\x_test_scores.csv",index=False)
