==> notebooks/09_Risk_Segmentation_and_Banding.ipynb <==
# Load the Train and test scores and targets

import pandas as pd
import numpy as np

train_scores=pd.read_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\x_train_scores.csv")
test_scores=pd.read_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\x_test_scores.csv")

y_train=pd.read_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\y_train.csv")
y_test=pd.read_csv(r"E:\DV Data Science\Retail_PD_Scorecard\data\processed\y_test.csv")

# Create the datasets with scores and target only

train_scoring=train_scores[["SCORE"]]

train_scoring["TARGET"]=y_train["TARGET"]

train_scoring.head()

test_scoring=test_scores[['SCORE']]

test_scoring['TARGET']=y_test['TARGET']

test_scoring.head()

# CREATE DECILES BASED ON SCORES FOR TRAIN

train_scoring["DECILES"], bin_edges=pd.qcut(train_scoring["SCORE"],q=10,retbins=True)

train_decile_summary=train_scoring.groupby("DECILES",as_index=False).agg(TOTAL=("SCORE","count"),BADS=("TARGET","sum"))

train_decile_summary

train_decile_summary["GOODS"]=train_decile_summary["TOTAL"]-train_decile_summary["BADS"]

train_decile_summary["BAD%"]=train_decile_summary["BADS"]/train_decile_summary["TOTAL"]
train_decile_summary["GOOD%"]=train_decile_summary["GOODS"]/train_decile_summary["TOTAL"]

train_decile_summary["CUM_BAD%"]=train_decile_summary["BADS"].cumsum()/train_decile_summary["BADS"].sum()
train_decile_summary["CUM_GOOD%"]=train_decile_summary["GOODS"].cumsum()/train_decile_summary["GOODS"].sum()

train_decile_summary

# CREATE A FUNCTION TO APPLY THE DECILES FUNCTION ON TEST

def get_decile_summary(dataframe,bin_edges):
    decile_summary=dataframe.copy(deep=True)
    decile_summary["DECILES"]=pd.cut(decile_summary["SCORE"],bins=bin_edges)
    decile_summary=decile_summary.groupby("DECILES", as_index=False).agg(TOTAL=("SCORE","count"),BADS=("TARGET","sum"))
    decile_summary["GOODS"]=decile_summary["TOTAL"]-decile_summary["BADS"]
    decile_summary["BADS%"]=decile_summary["BADS"]/decile_summary["TOTAL"]
    decile_summary["GOODS%"]=decile_summary["GOODS"]/decile_summary["TOTAL"]
    decile_summary["CUM_BAD%"]=decile_summary["BADS"].cumsum()/decile_summary["BADS"].sum()
    decile_summary["CUM_GOOD%"]=decile_summary["GOODS"].cumsum()/decile_summary["GOODS"].sum()
    return decile_summary

test_decile_summary=get_decile_summary(test_scoring,bin_edges)
test_decile_summary

train_scoring.head()

train_scoring["DECILES"], binned_edges=pd.qcut(train_scoring["SCORE"],q=10,labels=[0,1,2,3,4,5,6,7,8,9],retbins=True)
train_scoring.head()

grade_map = {
    0: "E", 1: "E",
    2: "D", 3: "D",
    4: "C", 5: "C",
    6: "B", 7: "B",
    8: "A", 9: "A"
}

train_scoring["GRADE"] = train_scoring["DECILES"].astype(int).map(grade_map)

train_scoring.head()

grade_summary_train=train_scoring.groupby("GRADE",as_index=False).agg(TOTAL=("TARGET","count"),BAD=("TARGET","sum"))

grade_summary_train["GOOD"]=grade_summary_train["TOTAL"]-grade_summary_train["BAD"]

grade_summary_train["BAD%"]=grade_summary_train["BAD"]/grade_summary_train["TOTAL"]
grade_summary_train["POPULATION%"]=grade_summary_train["TOTAL"]/grade_summary_train["TOTAL"].sum()

grade_summary_train

grade_summary_train["CUM_BAD%"]=grade_summary_train["BAD"].cumsum()/grade_summary_train["BAD"].sum()
grade_summary_train["CUM_POP%%"]=grade_summary_train["TOTAL"].cumsum()/grade_summary_train["TOTAL"].sum()

grade_summary_train

# Apply the same to test

test_scoring.head()

test_scoring["DECILES"]=pd.cut(test_scoring["SCORE"],bins=binned_edges,labels=[0,1,2,3,4,5,6,7,8,9])
test_scoring.head()

test_scoring["DECILES"].value_counts(dropna=False)

*2 test scores fall outside the train bin_edges range.*

#Replace first and last edges:
binned_edges[0] = -np.inf
binned_edges[-1] = np.inf

#Reapply pd.cut

test_scoring["DECILES"]=pd.cut(test_scoring["SCORE"],bins=binned_edges,labels=[0,1,2,3,4,5,6,7,8,9])
test_scoring.head()

test_scoring["GRADE"]=test_scoring["DECILES"].astype(int).map(grade_map)

test_scoring.head()

# Create grade-level summary for TEST

grade_summary_test = (
    test_scoring
        .groupby("GRADE", as_index=False)
        .agg(
            TOTAL=("SCORE", "count"),
            BAD=("TARGET", "sum")
        )
)

# Calculate GOODS
grade_summary_test["GOOD"] = grade_summary_test["TOTAL"] - grade_summary_test["BAD"]

# Bad rate
grade_summary_test["BAD%"] = grade_summary_test["BAD"] / grade_summary_test["TOTAL"]

# Population %
grade_summary_test["POPULATION%"] = grade_summary_test["TOTAL"] / grade_summary_test["TOTAL"].sum()

# Sort grades properly (A → E)
grade_summary_test = grade_summary_test.sort_values("GRADE").reset_index(drop=True)

grade_summary_test
